factor out runner implementation
